version test 1.16 Portfolio Screen partially implemented.

+++ app/src/main/java/com/example/myapplication/Cash.java
@@ -1,5 +1,5 @@
 package com.example.myapplication;
 
 public class Cash {
-    float balance;
+    double balance;
 }
